+++ source/asad_audio_encoder/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- added force encoding option
+- added check so file not reprocessed if stream encoder equals configured encoder
+
 **<span style="color:#56adda">0.0.2</span>**
 - Added new "Default/None" option which doesn't set a bitrate (and uses the codec default)
 - Changed codec names